Display exceptions as GUIs; #42
Depending on when in the process the exceptions occur, either return or exit
Add Shout to the Lord as a new song to test this functionality
